ai review

+++ albums-api/Models/CartModels.cs
@@ -1,9 +1,9 @@
 namespace albums_api.Models
 {
-    public record CartItem(int AlbumId, string Title, string Artist, double Price, string ImageUrl);
+    public record CartItem(int AlbumId, string Title, string Artist, decimal Price, string ImageUrl);
 
     public record Cart(string CartId, List<CartItem> Items)
     {
-        public double Total => Items.Sum(i => i.Price);
+        public decimal Total => Items.Sum(i => i.Price);
     }
 }
